+++ CHANGELOG.md
@@ -2,6 +2,11 @@
 
 ## Unreleased
 
+- Add support for Elasticsearch backend
+- Add configurable search synonyms
+- Use Porter stemming algorithm search filter, for English grammar
+- Add an English stop words search filter
+
 ## 28.0 (2020-11-25)
 
 Compare: https://git.torchbox.com/buckinghamshire-council/bc/compare/27.0...28.0